% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_to_low.R
\name{high_to_low}
\alias{high_to_low}
\title{Sort a Numeric Vector in Decreasing Order}
\usage{
high_to_low(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether to remove missing values
before sorting.}
}
\value{
A vector with the values sorted in descreasing order.
}
\description{
Sort a numeric vector so that the values are in deceasing order.  Missing
values are optionally removed or put last.
}
\examples{
x <- c(3, 7, 2, NA)
high_to_low(x)
high_to_low(x, na.rm = TRUE)
}
