% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_to_high.R
\name{low_to_high}
\alias{low_to_high}
\title{Sort a Numeric Vector in Increasing Order}
\usage{
low_to_high(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether to remove missing values
before sorting.}
}
\value{
A vector with the values sorted in increasing order.
}
\description{
\code{low_to_high} sorts the values of a numeric vector in
increasing order.
}
\details{
This is a convenience function that is a wrapper around
\code{\link{sort.int}}.
}
\examples{
x <- c(3, 7, 2, NA)
low_to_high(x)
low_to_high(x, na.rm = TRUE)
}
